Move multiply_mod_bnm1 to src/mulmod_bnm1.cpp
The wrapped product and its mod-(B^h + 1) helper are now compiled
once, taking only the type-erased scratch (whose heap hooks supply the
internal products via the new multiply_runtime_any and the cyclic NTT
tier's workspaces). The header keeps the constexpr fold helpers, the
size chooser, and the storage model; callers drop the allocator
argument.
