+++ CMakeLists.txt
@@ -40,6 +40,7 @@ add_library(
     beman.big_int
     src/karatsuba.cpp
     src/mul_dispatch.cpp
+    src/mulmod_bnm1.cpp
     src/toom_cook_3.cpp
     src/toom_cook_4.cpp
     src/toom_cook_6_5.cpp
